+++ config/config.php
@@ -1,7 +1,52 @@
 <?php
 
 return [
+    'eddbk' => [
+        'week_starts_on'        => 'sunday',
+        'default_calendar_view' => 'week',
+        'booking_wizard_color'  => '#17a7dd',
+        'booking_statuses_colors'     => [
+            'draft'     => '#dfe4ea',
+            'pending'   => '#1e90ff',
+            'scheduled' => '#2ed573',
+            'completed' => '#57606f',
+            'cancelled' => '#eb4d4b',
+        ],
+    ],
     'wp_bookings_ui' => [
+        'settings' => [
+            'options' => [
+                'week_starts_on' => [
+                    'sunday' => 'Sunday',
+                    'monday' => 'Monday',
+                    'tuesday' => 'Tuesday',
+                    'wednesday' => 'Wednesday',
+                    'thursday' => 'Thursday',
+                    'friday' => 'Friday',
+                    'saturday' => 'Saturday',
+                ],
+                'default_calendar_view' => [
+                    'day'   => 'Day',
+                    'week'  => 'Week',
+                    'month' => 'Month',
+                ]
+            ],
+            'fields' => [
+                'week_starts_on',
+                'default_calendar_view',
+                'booking_wizard_color',
+                'booking_statuses_colors',
+            ],
+            'prefix' => 'eddbk',
+            'array_fields' => [
+                'booking_statuses_colors'
+            ],
+            'action' => WP_BOOKINGS_UI_UPDATE_SETTINGS_ACTION,
+            'update_endpoint' => [
+                'method' => 'post',
+                'url' => admin_url('admin-ajax.php?action=' . WP_BOOKINGS_UI_UPDATE_SETTINGS_ACTION)
+            ],
+        ],
         'screen_options' => [
             'key' => WP_BOOKINGS_UI_SCREEN_OPTIONS_KEY,
             'fields' => [